Notifications and Refactoring Login/Signup Views
