+++ CalmCompass/ViewModels/AuthViewModel.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import Firebase
+import SwiftUI
 
 class AuthViewModel: ObservableObject {
     
@@ -19,21 +20,266 @@ class AuthViewModel: ObservableObject {
     
     //functions for handling authentication
     
-//    func login(){
-//        auth.signIn(withEmail: email, password: password) { result, error in
-//            if error != nil {
-//                print(error?.localizedDescription)
-//            }
-//        }
-//    }
-//    
-//    func register(){
-//        auth.createUser(withEmail: email, password: password){ result, error in
-//            if error != nil {
-//                print(error?.localizedDescription)
-//            }
-//            
-//        }
-//    }
+    struct getStartedView: View {
+        @State private var email = ""
+        @State private var password = ""
+        @State private var userIsLoggedIn = false
+        var auth = Auth.auth()
+        
+        var body: some View {
+            ZStack {
+                Color.white
+                
+                RoundedRectangle(cornerRadius: 30, style: . continuous)
+                    .foregroundStyle(.linearGradient(colors: [.white], startPoint: .topLeading, endPoint: .bottomTrailing))
+                    .frame(width:400, height:900)
+                
+                VStack(spacing: 20) {
+                    
+                    Image("startScreen")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 500, height: 300)
+                        .offset(y: -50)
+                        .overlay(
+                            Image("LetsLogin"))
+                        .scaledToFit()
+                        .frame(width: 400, height: 200)
+                    
+                    Text("Let us be your CalmCompass")
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .font(.system(size: 40, weight: .bold, design: . rounded))
+                        .offset(y: 20)
+                    
+                    NavigationLink(destination: SignedUpView()) {
+                        Text("Sign Up")
+                            .bold()
+                            .frame(width: 200, height: 40)
+                            .background(
+                                RoundedRectangle(cornerRadius: 10, style: .continuous)
+                                    .fill(.linearGradient(colors: [.purple, .blue], startPoint: .top, endPoint: .bottom))
+                            )
+                            .foregroundColor(.white)
+                    }
+                    .padding(.top)
+                    .offset(y: 100)
+                    
+                    NavigationLink(destination: inlogningView()) {
+                        Text("Already have an account? Login")
+                            .bold()
+                            .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                    }
+                    .padding(.top)
+                    .offset(y: 110)
+                    
+                }
+                .frame(width: 350)
+                .onAppear{
+                    Auth.auth().addStateDidChangeListener{auth, user in
+                        if user != nil {
+                            userIsLoggedIn.toggle()
+                        }
+                    }
+                }
+            }
+            .ignoresSafeArea()
+        }
+    }
+ 
+    struct inlogningView: View {
+        
+        @State private var email = ""
+        @State private var password = ""
+        @State private var navigateToHome = false
+        
+        var auth = Auth.auth()
+        
+        var body: some View {
+            ZStack {
+                Color.white
+                
+                RoundedRectangle(cornerRadius: 30, style: . continuous)
+                    .foregroundStyle(.linearGradient(colors: [.white], startPoint: .topLeading, endPoint: .bottomTrailing))
+                    .frame(width:400, height:900)
+                
+                VStack(spacing: 20) {
+                    Text("Welcome Back")
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .font(.system(size: 40, weight: .bold, design: . rounded))
+                        .offset(y: -70)
+                    
+                    Image("meditationLogin")
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 500, height: 300)
+                        .padding(.vertical, 10)
+                    
+                    TextField("Email", text: $email)
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .textFieldStyle(.plain)
+                        .placeholder(when: email.isEmpty) {
+                            Text("Email")
+                                .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                                .bold()
+                        }
+                    
+                    Rectangle()
+                        .frame(width: 350, height: 1)
+                        .foregroundColor(.white)
+                    
+                    SecureField("Password", text: $password)
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .textFieldStyle(.plain)
+                        .placeholder(when: password.isEmpty){
+                            Text("Password")
+                                .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                                .bold()
+                        }
+                    
+                    Rectangle()
+                        .frame(width: 350, height: 1)
+                        .foregroundColor(.white)
+                    
+                    Button {
+                        auth.signIn(withEmail: email, password: password) { result, error in
+                            if error != nil {
+                                print(error?.localizedDescription)
+                                //need navigation to HomeScreen after this
+                                
+                            }
+                        }
+                        navigateToHome = true
+                    } label: {
+                        Text("Login")
+                            .bold()
+                            .frame(width: 200, height: 40)
+                            .background(
+                                RoundedRectangle(cornerRadius: 10, style: .continuous)
+                                    .fill(.linearGradient(colors: [.purple, .blue], startPoint: .top, endPoint: .bottom))
+                            )
+                            .foregroundColor(.white)
+                    }
+                    .padding(.top)
+                    .offset(y: 10)
+                }
+                .frame(width: 350)
+            }
+            .ignoresSafeArea()
+            .navigationBarHidden(true)
+            .navigationTitle("")
+            .navigationBarTitleDisplayMode(.inline)
+            .background( NavigationLink(destination: HomeScreenView(), isActive: $navigateToHome) {
+                EmptyView()
+            })
+        }
+    }
+    
+    struct SignedUpView: View {
+        
+        @State private var email = ""
+        @State private var password = ""
+        @State private var username = ""
+        @State private var navigateToHome = false
+        
+        var auth = Auth.auth()
+        private let db = Firestore.firestore()
+        
+        var body: some View {
+            ZStack {
+                Color.white
+                
+                RoundedRectangle(cornerRadius: 30, style: . continuous)
+                    .foregroundStyle(.linearGradient(colors: [.white], startPoint: .topLeading, endPoint: .bottomTrailing))
+                    .frame(width:400, height:900)
+                
+                VStack(spacing: 20) {
+                    Text("  Welcome to CalmCompass")
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .font(.system(size: 40, weight: .bold, design: . rounded))
+                        .offset(y: -110)
+                    
+                    TextField("Email", text: $email)
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .textFieldStyle(.plain)
+                        .placeholder(when: email.isEmpty) {
+                            Text("Email")
+                                .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                                .bold()
+                        }
+                    
+                    Rectangle()
+                        .frame(width: 350, height: 1)
+                        .foregroundColor(.white)
+                    
+                    SecureField("Password", text: $password)
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .textFieldStyle(.plain)
+                        .placeholder(when: password.isEmpty){
+                            Text("Password")
+                                .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                                .bold()
+                        }
+                    
+                    Rectangle()
+                        .frame(width: 350, height: 1)
+                        .foregroundColor(.white)
+                    
+                    TextField("Name", text: $username)
+                        .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                        .textFieldStyle(.plain)
+                        .placeholder(when: username.isEmpty) {
+                            Text("Name")
+                                .foregroundStyle(.linearGradient(colors: [ .purple, .blue], startPoint: .topLeading, endPoint: .bottomTrailing))
+                                .bold()
+                        }
+                    
+                    Rectangle()
+                        .frame(width: 350, height: 1)
+                        .foregroundColor(.white)
+                    
+                    
+                    Button {
+                        auth.createUser(withEmail: email, password: password){ result, error in
+                            if let error = error {
+                                print(error.localizedDescription)
+                            }else if let user = result?.user {
+                                let userDocument = db.collection("users").document(user.uid)
+                                let userData = [
+                                    "username": username
+                                ]
+                                
+                                userDocument.setData(userData) {error in
+                                    if let error = error {
+                                        print("Error saving username: \(error.localizedDescription)")
+                                    } else {
+                                        print("Username saved successfully.")
+                                    }
+                                }
+                                //need navigation to HomeScreen after this
+                                navigateToHome = true
+                            }
+                        }
+                    } label: {
+                        Text("Sign Up")
+                            .bold()
+                            .frame(width: 200, height: 40)
+                            .background(
+                                RoundedRectangle(cornerRadius: 10, style: .continuous)
+                                    .fill(.linearGradient(colors: [.purple, .blue], startPoint: .top, endPoint: .bottom))
+                            )
+                            .foregroundColor(.white)
+                    }
+                    .padding(.top)
+                    .offset(y: 100)
+                }
+                .frame(width: 350)
+                NavigationLink(destination: HomeScreenView(), isActive: $navigateToHome) {
+                    EmptyView()
+                }
+            }
+            .ignoresSafeArea()
+        }
+    }
+    
 }
